Module-6/Demo-6/enterprise-mcp-weather/requirements.txt
mcp
fastmcp
requests
google-genai
python-dotenv
tenacity
streamlit